python/requirements.txt
PyQt5==5.15.11
python-can==4.5.0